+++ CLI/polymath-cli.js
@@ -15,7 +15,8 @@ const fs = require('fs');
 
 program
   .version('0.0.1')
-  .description('CLI for Polymath-core');
+  .description('CLI for Polymath-core')
+  .option('-r, --remote-node <network>', 'Use Infura to connect to a remote node on selected network');
 
 program
   .command('st20generator')
@@ -32,7 +33,7 @@ program
       mintingConfig = config.initialMint;
       stoCofig = config.sto;
     }
-    await st20generator.executeApp(tokenConfig, mintingConfig, stoCofig);
+    await st20generator.executeApp(tokenConfig, mintingConfig, stoCofig, program.remoteNode);
   });
 
 program
@@ -40,7 +41,7 @@ program
   .alias('f')
   .description('Poly faucet for local private netwtorks')
   .action(async function(beneficiary, amount) {
-    await faucet.executeApp(beneficiary, amount);
+    await faucet.executeApp(beneficiary, amount, program.remoteNode);
   });
 
 program
@@ -48,7 +49,7 @@ program
   .alias('i')
   .description('Participate in any STO you have been whitelisted for')
   .action(async function(investor, symbol, currency, amount) {
-    await investor_portal.executeApp(investor, symbol, currency, amount);
+    await investor_portal.executeApp(investor, symbol, currency, amount, program.remoteNode);
   });
 
 program
@@ -56,7 +57,7 @@ program
   .alias('mm')
   .description('View modules attached to a token and their status')
   .action(async function() {
-    await module_manager.executeApp();
+    await module_manager.executeApp(program.remoteNode);
   });
 
 program
@@ -64,7 +65,7 @@ program
   .alias('mi')
   .description('Distribute tokens to previously whitelisted investors')
   .action(async function(tokenSymbol, batchSize) {
-    shell.exec(`${__dirname}/commands/scripts/script.sh Multimint ${tokenSymbol} ${batchSize}`);;
+    shell.exec(`${__dirname}/commands/scripts/script.sh Multimint ${tokenSymbol} ${batchSize} ${program.remoteNode}`);;
   });
 
 program
@@ -72,7 +73,7 @@ program
   .alias('t')
   .description('Transfer ST tokens to another account')
   .action(async function(tokenSymbol, transferTo, transferAmount) {
-    await transfer.executeApp(tokenSymbol, transferTo, transferAmount);
+    await transfer.executeApp(tokenSymbol, transferTo, transferAmount, program.remoteNode);
   });
 
 program
@@ -80,7 +81,7 @@ program
   .alias('w')
   .description('Mass-update a whitelist of allowed/known investors')
   .action(async function(tokenSymbol, batchSize) {
-    shell.exec(`${__dirname}/commands/scripts/script.sh Whitelist ${tokenSymbol} ${batchSize}`);;
+    shell.exec(`${__dirname}/commands/scripts/script.sh Whitelist ${tokenSymbol} ${batchSize} ${remoteNetwork} ${program.remoteNode}`);
   });
 
 program
@@ -88,7 +89,7 @@ program
   .alias('erc20')
   .description('Runs erc20Explorer')
   .action(async function() {
-    await erc20explorer.executeApp();
+    await erc20explorer.executeApp(program.remoteNode);
   });
 
 program
@@ -96,7 +97,7 @@ program
   .alias('eth')
   .description('Runs ethExplorer')
   .action(async function() {
-    await ethExplorer.executeApp();
+    await ethExplorer.executeApp(program.remoteNode);
   });
 
 program
@@ -104,7 +105,15 @@ program
   .alias('a')
   .description('Runs accredit')
   .action(async function(tokenSymbol, batchSize) {
-    shell.exec(`${__dirname}/commands/scripts/script.sh Accredit ${tokenSymbol} ${batchSize}`);;
+    shell.exec(`${__dirname}/commands/scripts/script.sh Accredit ${tokenSymbol} ${batchSize} ${program.remoteNode}`);;
+  });
+
+program
+  .command('nonAccreditedLimit <tokenSymbol> [batchSize]')
+  .alias('nal')
+  .description('Runs changeNonAccreditedLimit')
+  .action(async function(tokenSymbol, batchSize) {
+    shell.exec(`${__dirname}/commands/scripts/script.sh NonAccreditedLimit ${tokenSymbol} ${batchSize} ${program.remoteNode}`);;
   });
 
 program
@@ -112,7 +121,12 @@ program
   .alias('str')
   .description('Runs STR Migrator')
   .action(async function(fromStrAddress, toStrAddress) {
-    await strMigrator.executeApp(fromStrAddress, toStrAddress);
+    await strMigrator.executeApp(fromStrAddress, toStrAddress, program.remoteNode);
   });
 
 program.parse(process.argv);
+
+if (typeof program.commands.length == 0) {
+  console.error('No command given!');
+  process.exit(1);
+}