+++ contracts/modules/Module.sol
@@ -57,7 +57,7 @@ contract Module is IModule {
      * @notice used to withdraw the fee by the factory owner
      */
     function takeFee(uint256 _amount) public withPerm(FEE_ADMIN) returns(bool) {
-        require(polyToken.transferFrom(address(this), Ownable(factory).owner(), _amount), "Unable to take fee");
+        require(polyToken.transferFrom(securityToken, Ownable(factory).owner(), _amount), "Unable to take fee");
         return true;
     }
 }