+++ contracts/modules/STO/USDTieredSTO.sol
@@ -22,6 +22,9 @@ contract USDTieredSTO is ISTO, ReentrancyGuard {
     string public ETH_ORACLE = "EthUsdOracle";
     mapping (bytes32 => mapping (bytes32 => string)) oracleKeys;
 
+    // Determine whether users can invest on behalf of a beneficiary
+    bool public allowBeneficialInvestments = false;
+
     // Address where ETH & POLY funds are delivered
     address public wallet;
 
@@ -70,8 +73,11 @@ contract USDTieredSTO is ISTO, ReentrancyGuard {
     // List of accredited investors
     mapping (address => bool) public accredited;
 
-    // Limit in USD for non-accredited investors multiplied by 10**18
+    // Default limit in USD for non-accredited investors multiplied by 10**18
     uint256 public nonAccreditedLimitUSD;
+
+    // Overrides for default limit in USD for non-accredited investors multiplied by 10**18
+    mapping (address => uint256) public nonAccreditedLimitUSDOverride;
 
     // Minimum investable amount in USD
     uint256 public minimumInvestmentUSD;
@@ -86,10 +92,14 @@ contract USDTieredSTO is ISTO, ReentrancyGuard {
     // Events //
     ////////////
 
+    event SetAllowBeneficialInvestments(bool _allowed);
+    event SetNonAccreditedLimit(address _investor, uint256 _limit);
+    event SetAccredited(address _investor, bool _accredited);
     event TokenPurchase(address indexed _purchaser, address indexed _beneficiary, uint256 _tokens, uint256 _usdAmount, uint256 _tierPrice, uint8 _tier);
     event FundsReceivedETH(address indexed _purchaser, address indexed _beneficiary, uint256 _usdAmount, uint256 _receivedValue, uint256 _spentValue, uint256 _rate);
     event FundsReceivedPOLY(address indexed _purchaser, address indexed _beneficiary, uint256 _usdAmount, uint256 _receivedValue, uint256 _spentValue, uint256 _rate);
     event ReserveTokenMint(address indexed _owner, address indexed _wallet, uint256 _tokens, uint8 _tier);
+
     event SetAddresses(
         address indexed _wallet,
         address indexed _reserveWallet
@@ -129,9 +139,7 @@ contract USDTieredSTO is ISTO, ReentrancyGuard {
     // STO Configuration //
     ///////////////////////
 
-    constructor (address _securityToken, address _polyAddress) public
-    Module(_securityToken, _polyAddress)
-    {      
+    constructor (address _securityToken, address _polyAddress) public Module(_securityToken, _polyAddress) {
         oracleKeys[bytes32("ETH")][bytes32("USD")] = ETH_ORACLE;
         oracleKeys[bytes32("POLY")][bytes32("USD")] = POLY_ORACLE;
     }
@@ -302,7 +310,33 @@ contract USDTieredSTO is ISTO, ReentrancyGuard {
         require(_investors.length == _accredited.length);
         for (uint256 i = 0; i < _investors.length; i++) {
             accredited[_investors[i]] = _accredited[i];
+            emit SetAccredited(_investors[i], _accredited[i]);
         }
+    }
+
+    /**
+     * @notice Modify the list of overrides for non-accredited limits in USD
+     * @param _investors Array of investor addresses to modify
+     * @param _nonAccreditedLimit Array of uints specifying non-accredited limits
+     */
+    function changeNonAccreditedLimit(address[] _investors, uint256[] _nonAccreditedLimit) public onlyOwner {
+        //nonAccreditedLimitUSDOverride
+        require(_investors.length == _nonAccreditedLimit.length);
+        for (uint256 i = 0; i < _investors.length; i++) {
+            require(_nonAccreditedLimit[i] > 0, "Limit cannot be 0");
+            nonAccreditedLimitUSDOverride[_investors[i]] = _nonAccreditedLimit[i];
+            emit SetNonAccreditedLimit(_investors[i], _nonAccreditedLimit[i]);
+        }
+    }
+
+    /**
+     * @notice Function to set allowBeneficialInvestments (allow beneficiary to be different to funder)
+     * @param _allowBeneficialInvestments Boolean to allow or disallow beneficial investments
+     */
+    function changeAllowBeneficialInvestments(bool _allowBeneficialInvestments) public onlyOwner {
+        require(_allowBeneficialInvestments != allowBeneficialInvestments, "Does not change value");
+        allowBeneficialInvestments = _allowBeneficialInvestments;
+        emit SetAllowBeneficialInvestments(allowBeneficialInvestments);
     }
 
     //////////////////////////
@@ -356,6 +390,9 @@ contract USDTieredSTO is ISTO, ReentrancyGuard {
       * @param _isPOLY Investment method
       */
     function _buyTokens(address _beneficiary, uint256 _investmentValue, uint256 _rate, bool _isPOLY) internal nonReentrant whenNotPaused returns(uint256, uint256) {
+        if (!allowBeneficialInvestments) {
+            require(_beneficiary == msg.sender, "Beneficiary must match funding provider");
+        }
         require(isOpen(), "STO is not open");
         require(_investmentValue > 0, "No funds were sent to buy tokens");
 
@@ -367,9 +404,10 @@ contract USDTieredSTO is ISTO, ReentrancyGuard {
 
         // Check for non-accredited cap
         if (!accredited[_beneficiary]) {
-            require(investorInvestedUSD[_beneficiary] < nonAccreditedLimitUSD, "Non-accredited investor has already reached nonAccreditedLimitUSD");
-            if (investedUSD.add(investorInvestedUSD[_beneficiary]) > nonAccreditedLimitUSD)
-                investedUSD = nonAccreditedLimitUSD.sub(investorInvestedUSD[_beneficiary]);
+            uint256 investorLimitUSD = (nonAccreditedLimitUSDOverride[_beneficiary] == 0) ? nonAccreditedLimitUSD : nonAccreditedLimitUSDOverride[_beneficiary];
+            require(investorInvestedUSD[_beneficiary] < investorLimitUSD, "Non-accredited investor has already reached nonAccreditedLimitUSD");
+            if (investedUSD.add(investorInvestedUSD[_beneficiary]) > investorLimitUSD)
+                investedUSD = investorLimitUSD.sub(investorInvestedUSD[_beneficiary]);
         }
 
         uint256 spentUSD;